add generic labelling strategy for dataset handler
