+++ ai/.gitignore
@@ -2,7 +2,9 @@ assets/
 test_assets/
 process/
 temp/
+temp.py
 __pycache__/
 ./**/__pycache__/
 .idea/
 venv/
+split_transient.py
